requirements.txt
mysql-connector-python~=8.0.29
requests~=2.25.1
progressbar~=2.5
pandas~=1.3.4